🎨 마이페이지 역할별 컴포넌트 분리
#39
